+++ test/index.test.ts
@@ -9,7 +9,8 @@ import { expect, describe, it, vi, afterEach, beforeEach } from "vitest";
 import M3U8Downloader from "../src/index.js";
 import { serverStart } from "./express.js";
 
-export const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
 
 describe("M3U8Downloader", () => {
   const m3u8Url = "http://127.0.0.1:3000/video.m3u8";
@@ -23,15 +24,77 @@ describe("M3U8Downloader", () => {
     const data = await axios.get("http://localhost:3000");
     expect(data.data).toEqual("Hello World!");
   });
-  it("should get the m3u8 content", async () => {
-    const downloader = new M3U8Downloader(m3u8Url, output) as any;
-    const data = await axios.get(m3u8Url);
-    const data2 = await downloader.getM3U8();
-    expect(data.data).toEqual(data2);
+  describe.concurrent("getM3U8", () => {
+    it("should get the m3u8 content", async () => {
+      const downloader = new M3U8Downloader(m3u8Url, output) as any;
+      const data = await axios.get(m3u8Url);
+      const data2 = await downloader.getM3U8();
+      expect(data.data).toEqual(data2);
+    });
+    it("should set the custom header", async () => {
+      const m3u8Url = "http://127.0.0.1:3000/head/video.m3u8";
+      const downloader = new M3U8Downloader(m3u8Url, output, {
+        headers: {
+          "custom-header": "custom-value",
+          "user-agent": "axios",
+        },
+      }) as any;
+      const data = await downloader.getM3U8();
+      expect(data["custom-header"]).toEqual("custom-value");
+      expect(data["user-agent"]).toEqual("axios");
+    });
+  });
+
+  describe("download", () => {
+    it("should download success", async () => {
+      const output = path.join(os.tmpdir(), "m3u8-downloader", "output.ts");
+
+      const downloader = new M3U8Downloader(m3u8Url, output, {
+        convert2Mp4: false,
+        tempDir,
+        // clean: false,
+      }) as any;
+
+      downloader.on("error", (error: string) => {
+        console.log("error", error);
+      });
+      downloader.on("progress", (progress: number) => {
+        console.log("progress", progress);
+      });
+      await downloader.download();
+      await sleep(100);
+
+      console.log("output", downloader.downloadedFiles);
+      expect(fs.existsSync(output)).toBeTruthy();
+      expect(fs.readFileSync(output).length).toEqual(8868524);
+
+      // clean
+      if (fs.existsSync(output)) {
+        fs.unlinkSync(output);
+      }
+    });
+    it("should download failure", async () => {
+      const m3u8Url = "http://127.0.0.1:3000/test.m3u8";
+      const output = path.join(os.tmpdir(), "m3u8-downloader", "output.ts");
+
+      const downloader = new M3U8Downloader(m3u8Url, output, {
+        convert2Mp4: false,
+        tempDir,
+      }) as any;
+
+      await downloader.download();
+      await sleep(100);
+
+      expect(downloader.status).toEqual("error");
+      expect(fs.existsSync(output)).toBeFalsy();
+      expect(fs.existsSync(path.join(tempDir, "segment0"))).toBeFalsy();
+      expect(fs.existsSync(path.join(tempDir, "segment1"))).toBeFalsy();
+    });
   });
 
   it("should parse the M3U8 file and return an array of URLs", () => {
     const downloader = new M3U8Downloader(m3u8Url, output) as any;
+    downloader.status = "running";
     const m3u8Content = "#EXTM3U\nsegment1.ts\nsegment2.ts";
 
     const tsUrls = downloader.parseM3U8(m3u8Content);
@@ -46,9 +109,10 @@ describe("M3U8Downloader", () => {
     const downloader = new M3U8Downloader(m3u8Url, output, {
       tempDir,
     }) as any;
+    downloader.status = "running";
     const tsUrls = [
-      "http://127.0.0.1:3000/segment01.ts",
-      "http://127.0.0.1:3000/segment02.ts",
+      "http://127.0.0.1:3000/segment0.ts",
+      "http://127.0.0.1:3000/segment1.ts",
     ];
 
     let downloadSegment = 0;
@@ -70,23 +134,23 @@ describe("M3U8Downloader", () => {
   });
 
   it("should merge the TS segments into a single file", async () => {
+    const tempDir = path.join(__dirname, "assets");
     const output = path.join(os.tmpdir(), "m3u8-downloader", "output.ts");
-    const tempDir = path.join(os.tmpdir(), "m3u8-downloader");
 
     const downloader = new M3U8Downloader(m3u8Url, output, {
       tempDir,
     }) as any;
-    const tsUrls = [
-      "http://127.0.0.1:3000/segment01.ts",
-      "http://127.0.0.1:3000/segment02.ts",
-    ];
+    downloader.status = "running";
 
-    const mergedFilePath = await downloader.mergeTsSegments(tsUrls, false);
+    const mergedFilePath = await downloader.mergeTsSegments(2, false);
     expect(mergedFilePath).toEqual(output);
+    await sleep(100);
     expect(fs.existsSync(output)).toBeTruthy();
+    expect(fs.readFileSync(output).length).toEqual(8868524);
 
     // clean
     if (fs.existsSync(output)) {
+      console.log("delete file", output);
       fs.unlinkSync(output);
     }
   });