+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # 0.2.0
 
-完善单元测试
+1. add unit test
+2. add mergeSegments option
+3. add clean option
+4. add headers option
+5. set default userAgent as "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:128.0) Gecko/20100101 Firefox/128.0"
+6. fix some bugs
 
 # 0.1.0